requirements.txt
pyyaml
pandas
requests
aiohttp
openpyxl